requirements.txt
streamlit
pandas
matplotlib
numpy
google-cloud-bigquery
google-cloud-bigquery-storage
google-auth
pyarrow